api/requirements.txt
fastapi>=0.115.14
pydantic>=2.11.7
python-dotenv==1.0.0
python-multipart==0.0.18
uvicorn>=0.34.2
langchain-openai>=0.3.27
langgraph>=0.5.1
tavily-python>=0.7.9
yfinance>=0.2.43
wikipedia==1.4.0